Add decorator to secure views requiring login

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -55,7 +56,6 @@ def all_products(request):
         'current_categories': categories,
         'current_sorting': current_sorting,
     }
-
     return render(request, 'products/products.html', context)
 
 
@@ -69,12 +69,16 @@ def product_detail(request, product_id):
     context = {
         'product': product,
     }
-
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product from the front end """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the owners can do that')
+        return redirect(reverse('home'))
+    
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,8 +98,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product from the front end """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the owners can do that')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -115,12 +124,16 @@ def edit_product(request, product_id):
         'form': form,
         'product': product,
     }
-
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the front end """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only the owners can do that')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product successfully deleted')
